Updated main level
Removed the BSP from under the main structure. Replaced with Board games
and proped up with some random items. Also (For testing) added a wire
from the cables plugin to see what it would looks like (On wall with the
poster) and added some lights sources to simulate fairly lights.